# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "SUSE AI Universal Proxy provides RESTful APIs for managing MCP (Model Context Protocol) server deployments and proxying requests to MCP servers running in Kubernetes.",
        "title": "SUSE AI Universal Proxy - Control Plane",
        "contact": {
            "name": "SUSE AI Universal Proxy Team",
            "url": "https://github.com/SUSE/suse-ai-up"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/SUSE/suse-ai-up/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8911",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/deployment/config/{serverId}": {
            "get": {
                "description": "Retrieve the configuration template for deploying an MCP server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment"
                ],
                "summary": "Get MCP server configuration template",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP Server ID",
                        "name": "serverId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MCPConfigTemplate"
                        }
                    },
                    "404": {
                        "description": "Server not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/deployment/deploy": {
            "post": {
                "description": "Deploy an MCP server to Kubernetes with provided configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "deployment"
                ],
                "summary": "Deploy an MCP server",
                "parameters": [
                    {
                        "description": "Deployment configuration",
                        "name": "deployment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DeployRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.DeployResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/v1/discovery/servers/{id}": {
            "get": {
                "description": "Returns a specific discovered MCP server by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discovery"
                ],
                "summary": "Get discovered server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DiscoveredServer"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.DeployRequest": {
            "type": "object",
            "required": [
                "server_id"
            ],
            "properties": {
                "env_vars": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "replicas": {
                    "type": "integer"
                },
                "resources": {
                    "$ref": "#/definitions/handlers.ResourceLimits"
                },
                "server_id": {
                    "type": "string"
                }
            }
        },
        "handlers.DeployResponse": {
            "type": "object",
            "properties": {
                "deployment_id": {
                    "type": "string"
                },
                "manifests": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "server_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ResourceLimits": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                }
            }
        },
...
        "models.ConnectionType": {
            "type": "string",
            "enum": [
                "SSE",
                "StreamableHttp",
                "RemoteHttp",
                "LocalStdio"
            ],
            "x-enum-varnames": [
                "ConnectionTypeSSE",
                "ConnectionTypeStreamableHttp",
                "ConnectionTypeRemoteHttp",
                "ConnectionTypeLocalStdio"
            ]
        },
        "models.DiscoveredServer": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "http://192.168.1.100:8000"
                },
                "connection": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ConnectionType"
                        }
                    ],
                    "example": "StreamableHttp"
                },
                "id": {
                    "type": "string",
                    "example": "server-123"
                },
                "lastSeen": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "MCP Example Server"
                },
                "protocol": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ServerProtocol"
                        }
                    ],
                    "example": "MCP"
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "vulnerability_score": {
                    "type": "string",
                    "example": "high"
                }
            }
        },
...
        "models.MCPConfigTemplate": {
            "type": "object",
            "properties": {
                "args": {
                    "description": "command arguments",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "command": {
                    "description": "docker, node, python, etc.",
                    "type": "string"
                },
                "env": {
                    "description": "environment variables",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "image": {
                    "description": "Docker image if applicable",
                    "type": "string"
                },
                "transport": {
                    "description": "stdio, http, sse, websocket",
                    "type": "string"
                }
            }
        },
...
        "models.ServerProtocol": {
            "type": "string",
            "enum": [
                "MCP"
            ],
            "x-enum-varnames": [
                "ServerProtocolMCP"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}